Comments work and are displayed on the map as well as on the unit page.

+++ OperationGlacier/Controllers/CommentsController.cs
@@ -8,12 +8,24 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using OperationGlacier.Models;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
 
 namespace OperationGlacier.Controllers
 {
     public class CommentsController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
+
+        /// <summary>
+        /// User manager - attached to application DB context
+        /// </summary>
+        protected UserManager<ApplicationUser> UserManager { get; set; }
+
+        public CommentsController()
+        {
+            this.UserManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(this.db));
+        }
 
         // GET: Comments
         public async Task<ActionResult> Index()
@@ -37,9 +49,37 @@ namespace OperationGlacier.Controllers
         }
 
         // GET: Comments/Create
-        public ActionResult Create()
+     //   public ActionResult Create()
+      //  {
+      //      return View();
+      //  }
+        /*
+         * 
+            + "date_string=" + current_unit.date_string
+            + "&unit_timeline_id=" + current_unit.timeline_id
+            + "&unit_location=" + current_unit.location
+            + "&x=" + current_unit.x
+            + "&y=" + current_unit.y + "'>Add Comment</a>";
+         * 
+         * */
+        public ActionResult Create(string date_string, string unit_timeline_id, string unit_location, int x, int y, string unit_name)
         {
-            return View();
+            if (!Request.IsAuthenticated)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            Comment model = new Comment()
+            {
+                date_in_world = DateTime.Now,
+                date_in_game = WitpUtility.from_date_str(date_string),
+                unit_timeline_id = unit_timeline_id,
+                x=x,
+                y=y,
+                unit_location = unit_location,
+                unit_name = unit_name
+            };
+            
+            return View(model);
         }
 
         // POST: Comments/Create
@@ -49,6 +89,14 @@ namespace OperationGlacier.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind(Include = "CommentID,Username,date_in_game,date_in_world,unit_timeline_id,unit_side_str,message,ReplyToCommentID,side_restriction,x,y,unit_name,unit_location,unit_report_first_line")] Comment comment)
         {
+            if (!Request.IsAuthenticated)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            comment.Username = user.UserName;
+            comment.side_restriction = user.SideRestriction;
+            comment.date_in_world = DateTime.Now;
             if (ModelState.IsValid)
             {
                 db.Comments.Add(comment);
@@ -62,6 +110,16 @@ namespace OperationGlacier.Controllers
         // GET: Comments/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (!Request.IsAuthenticated)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            if (user.UserName != "Baloogan")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -81,6 +139,16 @@ namespace OperationGlacier.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "CommentID,Username,date_in_game,date_in_world,unit_timeline_id,unit_side_str,message,ReplyToCommentID,side_restriction,x,y,unit_name,unit_location,unit_report_first_line")] Comment comment)
         {
+            if (!Request.IsAuthenticated)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            if (user.UserName != "Baloogan")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (ModelState.IsValid)
             {
                 db.Entry(comment).State = System.Data.Entity.EntityState.Modified;
@@ -93,6 +161,16 @@ namespace OperationGlacier.Controllers
         // GET: Comments/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (!Request.IsAuthenticated)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            if (user.UserName != "Baloogan")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -110,6 +188,16 @@ namespace OperationGlacier.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (!Request.IsAuthenticated)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            if (user.UserName != "Baloogan")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             Comment comment = await db.Comments.FindAsync(id);
             db.Comments.Remove(comment);
             await db.SaveChangesAsync();
